pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = ""
authors = ["R2 <volkovr_vk@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
stripe = "^10.8.0"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
